Add function to modify existing records (if needed)
This function will update the current stored beatmap data with the new data. If the new data is the same as the stored one, then nothing will be changed. If the new data is not the same as the stored one, the stored data will be change according to the new update data