Update stored API data updating function
The same change as last commit but to when user rrefresh the page again, the data can be updated automatically with the set mode type data aligned with the beatmap. If there is/are new changes to the current stored data, update it with the new records. Else, update the same values to the current record