Dynamic mod type assignment based on beatmap ID
New feature where it can dynamically assign mod types to beatmaps for every VOT tournament round. At the moment, it is only be used for the VOT4 Qualifiers round. This update includes a mapping of beatmap IDs to their respective mod types

The commit includes the following changes:
>> - Define an array of beatmap IDs for the VOT4 Qualifiers.
>> - Create a mapping of index ranges to corresponding mod types.
>> - Implement a function to retrieve the mod type by the array index.
>> - Update the data fetching loop to assign mod types based on the index number in the array.
>> - Insert or update beatmap data in the database with the assigned mod type.
>> - Retrieve and add beatmap data to an array for further processing.

TODO: we will try to make it re-usable for further round as well