Adjust fetch function for unauthenticated use-case
Previously, the function is hardset for fetching data from 2 database tables only. Now, it is set to dynamically fetch for data from its corresponding database tables follow by this keyword 'vot4_<sth>' for unauthenticated use-case. This is the preparation for the fully codes update to fetch data dynamically from all the database tables for both case without have to hardset it

Create another button for the 'Quarterfinals' round for the explained purpose above