Add new implement use-case for unauthenticed user
When user is not authenticated (I do not know why they do not want to but whatever), they can see the data fetched directly from the database table. However, this stored data will always be local data (a.k.a not up-to-date data)

To get up-to-date data, please just authenticated. It is very simple to do so. Thank you!