"feat(os-support): add dynamic CMake presets for Windows and Linux
I have decided to make it more dynamically for working on this project on both Linux and Windows (like I am right now). To be able to do so, I have make somequite big changes to be able to achieve that goal. Those are:

- Create a base 'CMakePresets.json' template file with Windows and Linux most minimal default configurations.
- Add an user-specifc CMake preset file to override and extend the base configuration for different build systems and operating systems. There will be a template file committing in later commit so that everyone know how to set it up correctly.
- Add conditional logic in 'CMakeLists.txt' file to handle different configurations based on the operating system.
- Including error handling for unsupported operating systems.
- Set global environment variables to idicate the operating systems will be used on current user's computer. There will be a documentation on how to do this correctly so everyone can set it up by themself.

BREAKING CHANGES: Because I aimed to make it very flexible on different OS, the whole files have pretty much refactorised so things will look a tiny bit different when building with this new template file. Please make sure to delete 'CMakeCache.txt' and 'CMakeFiles' directory everytime you want to build this project, especially please do so when using this new template on the first run. Thank you for your understanding!"